+++ src/Inshapardaz.Domain/Ports/Library/DeleteBookRequest.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using Inshapardaz.Domain.Repositories;
 using Inshapardaz.Domain.Repositories.Library;
 using Paramore.Brighter;
 
@@ -18,15 +19,32 @@ namespace Inshapardaz.Domain.Ports.Library
     public class DeleteBookRequestHandler : RequestHandlerAsync<DeleteBookRequest>
     {
         private readonly IBookRepository _bookRepository;
+        private readonly IFileRepository _fileRepository;
+        private readonly IFileStorage _fileStorage;
 
-        public DeleteBookRequestHandler(IBookRepository bookRepository)
+        public DeleteBookRequestHandler(IBookRepository bookRepository, IFileRepository fileRepository, IFileStorage fileStorage)
         {
             _bookRepository = bookRepository;
+            _fileRepository = fileRepository;
+            _fileStorage = fileStorage;
         }
 
         public override async Task<DeleteBookRequest> HandleAsync(DeleteBookRequest command, CancellationToken cancellationToken = new CancellationToken())
         {
-            await _bookRepository.DeleteBook(command.BookId, cancellationToken);
+            var author = await _bookRepository.GetBookById(command.BookId, cancellationToken);
+            if (author != null)
+            {
+                if (author.ImageId > 0)
+                {
+                    var image = await _fileRepository.GetFileById(author.ImageId, true, cancellationToken);
+                    if (image != null && !string.IsNullOrWhiteSpace(image.FilePath))
+                    {
+                        await _fileStorage.TryDeleteFile(image.FilePath, cancellationToken);
+                        await _fileRepository.DeleteFile(image.Id, cancellationToken);
+                    }
+                }
+                await _bookRepository.DeleteBook(command.BookId, cancellationToken);
+            }
 
             return await base.HandleAsync(command, cancellationToken);
         }
